feat: Write and pass tests for makeCurrentRecipe
