+++ src/data/data-model.js
@@ -1,6 +1,4 @@
 
-let currentRecipe = {};
-let currentRecipes = [];
-let savedRecipes = [];
 
-export { currentRecipe, currentRecipes, savedRecipes }
+
+// export { currentRecipe, currentRecipes, savedRecipes }
